tambah data product

+++ app/src/main/java/com/example/tugas1/model/Product.kt
@@ -5,7 +5,7 @@ import android.net.Uri
 data class Product(
     val id: String,
     var name: String,
-    var price: Int,
+    var price: Double,
     var description: String,
     var imageUrl: String? = null
 )
